+++ src/utils.rs
@@ -5,8 +5,8 @@ use crate::*;
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub struct Vec2 {
-    pub x: i32,
-    pub y: i32,
+    pub x: i8,
+    pub y: i8,
 }
 
 impl ops::Add<Vec2> for Vec2 {
@@ -28,8 +28,8 @@ pub fn set_draw_color(color: u16) {
 
 pub fn draw_cell(pos: Vec2) {
     rect(
-        pos.x * (CELL_SIZE as i32),
-        pos.y * (CELL_SIZE as i32),
+        (pos.x * (CELL_SIZE as i8)) as i32,
+        (pos.y * (CELL_SIZE as i8)) as i32,
         CELL_SIZE,
         CELL_SIZE,
     );
